♻️  cargo: Move RustEdition into own mod
This will be used outside package now as well.
