+++ tests/simple.rs
@@ -81,9 +81,7 @@ fn simple_cargo_toml() {
 #[cfg(feature = "cargo-toml")]
 #[test]
 fn simple_cargo_toml_serde() {
-    use tomling::cargo::{
-        package::RustEdition, BuildDependency, Dependency, Manifest, ResolverVersion,
-    };
+    use tomling::cargo::{BuildDependency, Dependency, Manifest, ResolverVersion, RustEdition};
 
     let manifest: Manifest = tomling::from_str(CARGO_TOML).unwrap();
