+++ tests/tokio.rs
@@ -65,7 +65,7 @@ fn tokio() {
 #[cfg(feature = "cargo-toml")]
 #[test]
 fn tokio_serde() {
-    use tomling::cargo::{package::RustEdition, Dependency, DevDependency, Manifest};
+    use tomling::cargo::{Dependency, DevDependency, Manifest, RustEdition};
 
     let manifest: Manifest = tomling::from_str(CARGO_TOML).unwrap();
     assert_eq!(manifest.package().name(), "tokio");
